+++ Modules/Frontend/resources/views/components/sections/top-pict.blade.php
@@ -1,6 +1,6 @@
 <div class="top-pics-block section-wraper">
     <div class="d-flex align-items-center justify-content-between px-1 mb-2 pb-1 mb-md-4 pb-md-0">
-        <h4 class="main-title text-capitalize mb-0 fw-medium">{{ __('sectionTitle.top_picks') }}</h4>
+        <h4 class="main-title text-capitalize mb-0 fw-medium">{{ $sectionHeading ?? __('sectionTitle.top_picks') }}</h4>
         <a href="{{ route('frontend.rail_archive', 'top-picks') }}" class="text-primary iq-view-all text-decoration-none flex-none">{{ __('streamButtons.view_all') }}</a>
     </div>
     <div class="card-style-slider">